Refactor endpoints into separate resources
Group related endpoints in a separate resource and mount then under
a common prefix.
